Refactor parser/predictive package
